Arrumado tipo de dado da struct enderecos_bitcoin

+++ main.c
@@ -18,34 +18,40 @@
 #define DATA_LENGTH 184
 #define RECOMPENSA_MINERACAO 50
 
-struct bloco_nao_minerado {
+struct bloco_nao_minerado
+{
   uint32_t numero;
   uint32_t nonce;
   unsigned char data[DATA_LENGTH];
   unsigned char hash_anterior[SHA256_DIGEST_LENGTH];
 };
 
-struct bloco_minerado {
+struct bloco_minerado
+{
   struct bloco_nao_minerado bloco;
   unsigned char hash[SHA256_DIGEST_LENGTH];
   struct bloco_minerado *prox;
 };
 
-struct sistema_bitcoin {
+struct sistema_bitcoin
+{
   uint32_t carteira[NUM_ENDERECOS];
 };
 
-struct enderecos_bitcoin {
-  uint32_t chave;
+struct enderecos_bitcoin
+{
+  uint8_t chave;
   struct enderecos_bitcoin *prox;
 };
 
-void iniciar_carteira(struct sistema_bitcoin *sistema) {
+void iniciar_carteira(struct sistema_bitcoin *sistema)
+{
   memset(sistema->carteira, 0, sizeof(sistema->carteira));
 }
 
 void inserir_enderecos_com_bitcoins(struct enderecos_bitcoin **raiz,
-                                    uint32_t chave) {
+                                    uint8_t chave)
+{
   struct enderecos_bitcoin *novo =
       (struct enderecos_bitcoin *)malloc(sizeof(struct enderecos_bitcoin));
 
@@ -56,20 +62,25 @@ void inserir_enderecos_com_bitcoins(struct enderecos_bitcoin **raiz,
   *raiz = novo;
 }
 
-void encontrar_maior_numero_bitcoins(uint32_t carteira[]) {
+void encontrar_maior_numero_bitcoins(uint32_t carteira[])
+{
   int num_enderecos_max_bitcoins = 0;
   uint32_t max_bitcoins = 0;
 
-  for (size_t i = 0; i < NUM_ENDERECOS; i++) {
-    if (carteira[i] > max_bitcoins) {
+  for (size_t i = 0; i < NUM_ENDERECOS; i++)
+  {
+    if (carteira[i] > max_bitcoins)
+    {
       max_bitcoins = carteira[i];
     }
   }
 
   uint8_t *enderecos_max_bitcoins = malloc(sizeof(uint8_t));
   assert(enderecos_max_bitcoins != NULL);
-  for (size_t i = 0; i < NUM_ENDERECOS; i++) {
-    if (carteira[i] == max_bitcoins) {
+  for (size_t i = 0; i < NUM_ENDERECOS; i++)
+  {
+    if (carteira[i] == max_bitcoins)
+    {
       enderecos_max_bitcoins[num_enderecos_max_bitcoins] = i;
       num_enderecos_max_bitcoins++;
 
@@ -82,32 +93,38 @@ void encontrar_maior_numero_bitcoins(uint32_t carteira[]) {
 
   printf("Maior numero de bitcoins = %u nos seguintes enderecos\n",
          max_bitcoins);
-  for (size_t i = 0; i < num_enderecos_max_bitcoins; i++) {
+  for (size_t i = 0; i < num_enderecos_max_bitcoins; i++)
+  {
     printf("%u\n", enderecos_max_bitcoins[i]);
   }
 }
 
-int escolhe_carteira(struct enderecos_bitcoin *raiz, uint32_t indice) {
+int escolhe_carteira(struct enderecos_bitcoin *raiz, uint8_t indice)
+{
   size_t i = 0;
   if (!raiz)
     return 0;
-  else if (raiz) {
-    while (i != indice && raiz != NULL) {
+  else if (raiz)
+  {
+    while (i != indice && raiz != NULL)
+    {
       raiz = raiz->prox;
       i++;
     }
     return raiz->chave;
-
-  } else
+  }
+  else
     return 0;
 }
 
-int conta_enderecos(struct enderecos_bitcoin *raiz) {
+int conta_enderecos(struct enderecos_bitcoin *raiz)
+{
   size_t cont = 0;
   if (!raiz)
     return 0;
   if (raiz)
-    while (raiz) {
+    while (raiz)
+    {
       raiz = raiz->prox;
       cont++;
     }
@@ -117,11 +134,13 @@ int conta_enderecos(struct enderecos_bitcoin *raiz) {
 void gerar_transacoes_bloco(struct bloco_nao_minerado *bloco,
                             unsigned long int endereco_minerador,
                             struct sistema_bitcoin *sistema, MTRand rand,
-                            struct enderecos_bitcoin **raiz) {
+                            struct enderecos_bitcoin **raiz)
+{
   size_t num_transacoes = 0;
   uint32_t carteira_auxiliar[NUM_ENDERECOS] = {0};
 
-  while (num_transacoes < MAX_TRANSACOES_BLOCO) {
+  while (num_transacoes < MAX_TRANSACOES_BLOCO)
+  {
     size_t num_enderecos_com_bitcoins = 0;
 
     int valor_lista_origem = genRandLong(&rand) % (conta_enderecos(*raiz));
@@ -131,7 +150,8 @@ void gerar_transacoes_bloco(struct bloco_nao_minerado *bloco,
 
     unsigned long int endereco_destino;
 
-    do {
+    do
+    {
       endereco_destino = genRandLong(&rand) % NUM_ENDERECOS;
     } while (endereco_origem == endereco_destino);
 
@@ -150,15 +170,18 @@ void gerar_transacoes_bloco(struct bloco_nao_minerado *bloco,
 
   bloco->data[DATA_LENGTH - 1] = (unsigned char)endereco_minerador;
 
-  for (size_t i = 0; i < NUM_ENDERECOS; i++) {
-    if (carteira_auxiliar[i] > 0) {
+  for (size_t i = 0; i < NUM_ENDERECOS; i++)
+  {
+    if (carteira_auxiliar[i] > 0)
+    {
       sistema->carteira[i] += carteira_auxiliar[i];
       inserir_enderecos_com_bitcoins(raiz, i);
     }
   }
 }
 
-void minerar_bloco(struct bloco_nao_minerado *bloco, unsigned char *hash) {
+void minerar_bloco(struct bloco_nao_minerado *bloco, unsigned char *hash)
+{
   uint32_t nonce = 0;
 
   unsigned char
@@ -166,7 +189,8 @@ void minerar_bloco(struct bloco_nao_minerado *bloco, unsigned char *hash) {
 
   unsigned char hash_result[SHA256_DIGEST_LENGTH];
 
-  while (nonce < UINT32_MAX) {
+  while (nonce < UINT32_MAX)
+  {
     memcpy(bloco_completo, bloco, sizeof(struct bloco_nao_minerado));
 
     memcpy(bloco_completo + sizeof(struct bloco_nao_minerado), &nonce,
@@ -175,7 +199,8 @@ void minerar_bloco(struct bloco_nao_minerado *bloco, unsigned char *hash) {
     SHA256(bloco_completo, sizeof(struct bloco_nao_minerado) + sizeof(uint32_t),
            hash_result);
 
-    if (hash_result[0] == 0 && hash_result[1] == 0) {
+    if (hash_result[0] == 0 && hash_result[1] == 0)
+    {
       memcpy(hash, hash_result, SHA256_DIGEST_LENGTH);
       bloco->nonce = nonce;
       break;
@@ -186,7 +211,8 @@ void minerar_bloco(struct bloco_nao_minerado *bloco, unsigned char *hash) {
 }
 
 void inserir_bloco(struct bloco_minerado **blockchain,
-                   struct bloco_nao_minerado bloco, unsigned char *hash) {
+                   struct bloco_nao_minerado bloco, unsigned char *hash)
+{
 
   struct bloco_minerado *novo_bloco =
       (struct bloco_minerado *)malloc(sizeof(struct bloco_minerado));
@@ -197,11 +223,15 @@ void inserir_bloco(struct bloco_minerado **blockchain,
   memcpy(novo_bloco->hash, hash, SHA256_DIGEST_LENGTH);
   novo_bloco->prox = NULL;
 
-  if (*blockchain == NULL) {
+  if (*blockchain == NULL)
+  {
     *blockchain = novo_bloco;
-  } else {
+  }
+  else
+  {
     struct bloco_minerado *atual = *blockchain;
-    while (atual->prox != NULL) {
+    while (atual->prox != NULL)
+    {
       atual = atual->prox;
     }
     atual->prox = novo_bloco;
@@ -209,10 +239,12 @@ void inserir_bloco(struct bloco_minerado **blockchain,
 }
 
 void imprimir_blocos_minerados(struct bloco_minerado *blockchain,
-                               struct sistema_bitcoin sistema) {
+                               struct sistema_bitcoin sistema)
+{
   struct bloco_minerado *atual = blockchain;
 
-  while (atual != NULL) {
+  while (atual != NULL)
+  {
     printf("Bloco %u:\n", atual->bloco.numero);
     printf("Hash do bloco: ");
 
@@ -227,9 +259,11 @@ void imprimir_blocos_minerados(struct bloco_minerado *blockchain,
 
     printf("Endereço do minerador: %u\n", atual->bloco.data[DATA_LENGTH - 1]);
 
-    if (atual->bloco.numero != 1) {
+    if (atual->bloco.numero != 1)
+    {
 
-      for (int i = 0; i < MAX_TRANSACOES_BLOCO; i++) {
+      for (int i = 0; i < MAX_TRANSACOES_BLOCO; i++)
+      {
         uint32_t endereco_origem = atual->bloco.data[i * 3];
         uint32_t endereco_destino = atual->bloco.data[i * 3 + 1];
         uint32_t quantidade = atual->bloco.data[i * 3 + 2];
@@ -247,11 +281,13 @@ void imprimir_blocos_minerados(struct bloco_minerado *blockchain,
   }
 }
 
-void remove_no(struct enderecos_bitcoin **raiz, uint32_t chave) {
+void remove_no(struct enderecos_bitcoin **raiz, uint8_t chave)
+{
   if (*raiz == NULL)
     return;
 
-  if ((*raiz)->chave == chave) {
+  if ((*raiz)->chave == chave)
+  {
     struct enderecos_bitcoin *temp = *raiz;
 
     *raiz = (*raiz)->prox;
@@ -266,7 +302,8 @@ void remove_no(struct enderecos_bitcoin **raiz, uint32_t chave) {
   struct enderecos_bitcoin *atual = *raiz;
   struct enderecos_bitcoin *anterior = NULL;
 
-  while (atual != NULL && atual->chave != chave) {
+  while (atual != NULL && atual->chave != chave)
+  {
     anterior = atual;
     atual = atual->prox;
   }
@@ -278,15 +315,18 @@ void remove_no(struct enderecos_bitcoin **raiz, uint32_t chave) {
 }
 
 int checa_zero_bitcoin(struct enderecos_bitcoin **raiz,
-                       struct sistema_bitcoin *sistema) {
+                       struct sistema_bitcoin *sistema)
+{
 
   if (*raiz == NULL)
     return 1;
   struct enderecos_bitcoin *atual = *raiz;
 
-  while (atual) {
+  while (atual)
+  {
 
-    if (sistema->carteira[atual->chave] <= 0) {
+    if (sistema->carteira[atual->chave] <= 0)
+    {
       // printf("Remoção do %d\n", atual->chave);
       // printf("Valor na carteira: %d\n", sistema->carteira[140]);
       remove_no(raiz, atual->chave);
@@ -297,8 +337,10 @@ int checa_zero_bitcoin(struct enderecos_bitcoin **raiz,
   return 1;
 }
 
-void imprime_lista(struct enderecos_bitcoin *raiz) {
-  while (raiz != NULL) {
+void imprime_lista(struct enderecos_bitcoin *raiz)
+{
+  while (raiz != NULL)
+  {
     printf("%d ",
            raiz->chave); // Supondo que o campo a ser impresso seja 'valor'
     raiz = raiz->prox;
@@ -307,8 +349,10 @@ void imprime_lista(struct enderecos_bitcoin *raiz) {
   printf("\n");
 }
 
-void free_blockchain(struct bloco_minerado **blockchain) {
-  while (*blockchain != NULL) {
+void free_blockchain(struct bloco_minerado **blockchain)
+{
+  while (*blockchain != NULL)
+  {
     struct bloco_minerado *proximo = (*blockchain)->prox;
     free(*blockchain);
     *blockchain = proximo;
@@ -317,7 +361,8 @@ void free_blockchain(struct bloco_minerado **blockchain) {
 
 void processar_bloco(struct enderecos_bitcoin **raiz,
                      struct bloco_minerado **blockchain,
-                     struct sistema_bitcoin *sistema) {
+                     struct sistema_bitcoin *sistema)
+{
 
   struct bloco_nao_minerado bloco;
 
@@ -327,13 +372,15 @@ void processar_bloco(struct enderecos_bitcoin **raiz,
   unsigned char hash_anterior[SHA256_DIGEST_LENGTH] = {0};
   MTRand rand = seedRand(1234567);
 
-  for (numero_bloco = 1; numero_bloco <= TOTAL_BLOCOS; ++numero_bloco) {
+  for (numero_bloco = 1; numero_bloco <= TOTAL_BLOCOS; ++numero_bloco)
+  {
     bloco.numero = numero_bloco;
     memcpy(bloco.hash_anterior, hash_anterior, SHA256_DIGEST_LENGTH);
 
     unsigned char hash[SHA256_DIGEST_LENGTH];
 
-    if (numero_bloco == 1) {
+    if (numero_bloco == 1)
+    {
       const char *mensagem = "The Times 03/Jan/2009 Chancellor on brink of "
                              "second bailout for banks";
 
@@ -351,7 +398,9 @@ void processar_bloco(struct enderecos_bitcoin **raiz,
 
       sistema->carteira[(uint32_t)bloco.data[DATA_LENGTH - 1]] +=
           RECOMPENSA_MINERACAO;
-    } else {
+    }
+    else
+    {
 
       unsigned long int endereco_minerador = genRandLong(&rand) % NUM_ENDERECOS;
 
@@ -375,7 +424,8 @@ void processar_bloco(struct enderecos_bitcoin **raiz,
   }
 }
 
-int main() {
+int main()
+{
 
   struct enderecos_bitcoin *raiz = NULL;
   struct bloco_minerado *blockchain = NULL;
